+++ src/configToSchemas/commonPlugins/base.js
@@ -1,6 +1,5 @@
 export default (config) => (props) => {
   props.schema ??= {};
-  props.schema.required ??= config.required;
   props.schema.readOnly ??= config.readOnly;
 
   props.uiSchema ??= {};